Move classes to the correct file
In the next commit this function will be used not only for RuleSets
