+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/rules/RuleSets.kt
@@ -1,18 +1,10 @@
 package io.gitlab.arturbosch.detekt.core.rules
 
 import io.github.detekt.tooling.api.spec.RulesSpec
-import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.Rule
 import io.gitlab.arturbosch.detekt.api.RuleSet
 import io.gitlab.arturbosch.detekt.api.RuleSetProvider
-import io.gitlab.arturbosch.detekt.api.internal.createPathFilters
 import io.gitlab.arturbosch.detekt.core.ProcessingSettings
-import org.jetbrains.kotlin.psi.KtFile
-
-fun Config.shouldAnalyzeFile(file: KtFile): Boolean {
-    val filters = createPathFilters()
-    return filters == null || !filters.isIgnored(file)
-}
 
 fun associateRuleIdsToRuleSetIds(ruleSets: List<RuleSet>): Map<Rule.Id, RuleSet.Id> {
     return ruleSets
