+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/util/ConfigExtensions.kt
@@ -1,5 +1,12 @@
 package io.gitlab.arturbosch.detekt.core.util
 
 import io.gitlab.arturbosch.detekt.api.Config
+import io.gitlab.arturbosch.detekt.api.internal.createPathFilters
+import org.jetbrains.kotlin.psi.KtFile
 
 internal fun Config.isActiveOrDefault(default: Boolean): Boolean = valueOrDefault(Config.ACTIVE_KEY, default)
+
+internal fun Config.shouldAnalyzeFile(file: KtFile): Boolean {
+    val filters = createPathFilters()
+    return filters == null || !filters.isIgnored(file)
+}
